requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
idna==3.10
joblib==1.5.1
nltk==3.9.1
oauthlib==3.2.2
python-dotenv==1.1.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
textblob==0.19.0
tqdm==4.67.1
tweepy==4.15.0
urllib3==2.4.0
